+++ packages/utils/src/index.ts
@@ -1,5 +1,6 @@
 export * from './currency.js'
 export * from './deals-csv.js'
+export * from './payments-csv.js'
 export * from './date.js'
 export * from './html.js'
 export * from './payment-invoice-html.js'